+++ pkg/proc/heap_test.go
@@ -37,3 +37,65 @@ func TestHeapBits(t *testing.T) {
 		}
 	}
 }
+
+func TestSmallSpanBitmapBase(t *testing.T) {
+	spanBase := Address(0x1000)
+	spanSize := int64(0x1000)
+	inlineSize := int64(128)
+	span := &spanInfo{base: spanBase, spanSize: spanSize}
+	bitmapSize := spanSize / 8 / 8
+	legacyBase := spanBase.Add(spanSize - bitmapSize)
+	tests := []struct {
+		name            string
+		greenTeaEnabled bool
+		hasInlineBitmap bool
+		inlineMarkSize  int64
+		want            Address
+	}{
+		{
+			name:            "legacy layout when greentea disabled",
+			greenTeaEnabled: false,
+			hasInlineBitmap: true,
+			inlineMarkSize:  inlineSize,
+			want:            legacyBase,
+		},
+		{
+			name:            "green tea inline mark bits shifts bitmap base",
+			greenTeaEnabled: true,
+			hasInlineBitmap: true,
+			inlineMarkSize:  inlineSize,
+			want:            legacyBase.Add(-inlineSize),
+		},
+		{
+			name:            "inline mark bits with zero inline size fallback",
+			greenTeaEnabled: true,
+			hasInlineBitmap: true,
+			inlineMarkSize:  0,
+			want:            legacyBase,
+		},
+		{
+			name:            "greentea enabled without inline flag uses legacy base",
+			greenTeaEnabled: true,
+			hasInlineBitmap: false,
+			inlineMarkSize:  inlineSize,
+			want:            legacyBase,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			scope := &HeapScope{
+				pageSize:               spanSize,
+				greenTeaGCEnabled:      tc.greenTeaEnabled,
+				spanInlineMarkBitsSize: tc.inlineMarkSize,
+			}
+			if tc.hasInlineBitmap {
+				scope.inlineMarkSpanPages = map[Address]struct{}{spanBase: {}}
+			}
+			got := smallSpanBitmapBase(scope, span)
+			if got != tc.want {
+				t.Fatalf("smallSpanBitmapBase = %#x, want %#x", uint64(got), uint64(tc.want))
+			}
+		})
+	}
+}
