+++ test/greenteagc_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build go1.26 && goexperiment.greenteagc
+//go:build goexperiment.greenteagc
 
 package test
